updated user invite audit log return data refactor clients updated
