+++ source/Energinet.DataHub.MarketParticipant.Application/Commands/User/GetUserAuditLogResponse.cs
@@ -13,10 +13,9 @@
 // limitations under the License.
 
 using System.Collections.Generic;
-using Energinet.DataHub.MarketParticipant.Domain.Model.Users;
 
 namespace Energinet.DataHub.MarketParticipant.Application.Commands.User;
 
 public sealed record GetUserAuditLogResponse(
     IEnumerable<UserRoleAssignmentAuditLogEntryDto> UserRoleAssignmentAuditLogs,
-    IEnumerable<UserInviteDetailsAuditLog> InviteDetailsAuditLogs);
+    IEnumerable<UserInviteDetailsAuditLogDto> InviteDetailsAuditLogs);
